Create a separate test file
